Move all apps projects to the top level src directory
A simple rebar include of the erlangsp library does not work if the
projects are not in the top level src directory. Rebar will not descend
to the apps directory when looking for a depenedency project's *.app.src
file. Since the erlangsp and coop projects should be directly included
as included_applications in a separate git project's *.app.src file,
everything is re-rooted to erlangsp/src.